refactor(page): extract usePrintingPreviewManager hook (phase 3)
Move the printing layer-preview interactive subsystem (layer scrub/zoom/
pan state + refs, wheel/pointer handlers, settle/PNG-load/settled-canvas
effects, transform/cursor memos) out of page.tsx into
features/printing/usePrintingPreviewManager.ts. printingPreviewTarget-
Resolution stays in Home (derives from profile + slicing artifact) and is
injected via a deps-ref; everything else is destructure-same-names so the
<PrintingPreviewPane> call + LayerSlider + scene scrubbing are unchanged.
56 names returned. Worktree subagent, integrated + re-verified here.

page.tsx: 16,775 -> 16,302 lines (-473). Verified via `tsc --noEmit`.
Smoke-test printing-mode layer scrub / zoom / pan before merge.
